Rename code example stimulus controller
