+++ app/helpers/code_helper.rb
@@ -7,7 +7,7 @@ module CodeHelper
     data = {language: language}
 
     if options[:runnable]
-      data[:controller] = "code"
+      data[:controller] = "code-example"
     end
 
     tag.div(
@@ -27,7 +27,7 @@ module CodeHelper
 
       body = tag.div(class: "code-body") do
         tag.pre do
-          tag.code data: {code_target: "source"} do
+          tag.code data: {code_example_target: "source"} do
             raw code_formatter.format(lexer.lex(code))
           end
         end + clipboard_copy(code)
@@ -37,13 +37,13 @@ module CodeHelper
       if options[:runnable]
         footer = tag.div(class: "code-footer") do
           tag.div(class: "code-actions") do
-            tag.button("Run", class: "button primary", data: {action: "click->code#run"}) +
-              tag.span(class: "code-action-status", data: {code_target: "status"})
+            tag.button("Run", class: "button primary", data: {action: "click->code-example#run"}) +
+              tag.span(class: "code-action-status", data: {code_example_target: "status"})
           end +
-            tag.pre(class: "code-output hidden", data: {code_target: "output"}) do
+            tag.pre(class: "code-output hidden", data: {code_example_target: "output"}) do
               tag.code
             end +
-            tag.pre(class: "code-result hidden", data: {code_target: "result"}) do
+            tag.pre(class: "code-result hidden", data: {code_example_target: "result"}) do
               tag.code
             end
         end
