+++ app/javascript/controllers/code-example/rails.js
