+++ app/javascript/controllers/index.js
@@ -4,19 +4,22 @@
 
 import { application } from './application';
 
-import Code from './code';
+import CodeExample from './code-example';
 import ClipboardCopy from './clipboard-copy';
 import Darkmode from './darkmode';
-import TableOfContents from './table-of-contents';
 
 import PwaInstallation from './pwa/installation';
 import PwaWebPushSubscription from './pwa/web-push-subscription';
 import PwaWebPushDemo from './pwa/web-push-demo';
 
-application.register('code', Code);
+import TableOfContents from './table-of-contents';
+
+application.register('code-example', CodeExample);
 application.register('clipboard-copy', ClipboardCopy);
 application.register('darkmode', Darkmode);
-application.register('table-of-contents', TableOfContents);
+
 application.register('pwa-installation', PwaInstallation);
 application.register('pwa-web-push-subscription', PwaWebPushSubscription);
 application.register('pwa-web-push-demo', PwaWebPushDemo);
+
+application.register('table-of-contents', TableOfContents);
